Auto Suggestor And Predictor
This is a research and development type project where we research about
the low speed or Urdu typing and high demand of it that make it more
expensive and less available in your surroundings so we found out that
due to 35+ alphabets in Urdu language, the international ISO standard
keyboards are only on English alphabets that are 25+ that make a quiet
big difference of about 10 alphabets that means we have to press and
hold SHIFT key while typing these 10+ alphabets that is wasting our time
and slowing our speed of typing so we tried to solved this problem by
keeping the standard along as they are. We made some programs to let you
type less and choose more. Here we are suggesting the word without
intelligence that what you want before your typing the full words and
also predicting the next word to be typed with our intelligence and
running many algorithm over it to make the probability closer to your
demand. Also, we are not fixing our idea on any fix standard as we
believe in that not every user is connected to other one so we made it
domain free that will make it more intelligent under user and user will
feel more comfortable with suggestion and predictions as per there
demand. So, this project help the typist to type more fast and less
pressing keys in Urdu typing and we cover over 10Billions devices by our
cross-platform codes of every type compatibility.